Modify route logic
