+++ src/main.jsx
@@ -1,12 +1,33 @@
 import ReactDOM from 'react-dom/client'
-import { CssBaseline, ThemeProvider } from '@mui/material'
 
-import { App } from './App'
+import { App } from './pages/App'
+import { ErrorPage } from './components/ErrorPage'
+import { Pokemon } from './components/Pokemon'
+
+import { StyledEngineProvider } from '@mui/material/styles'
+import { CssBaseline, ThemeProvider } from '@mui/material'
 import { theme } from './utils/theme'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom'
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        path: 'pokemon/:name',
+        element: <Pokemon />
+      }
+    ]
+  }
+])
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <ThemeProvider theme={theme}>
-    <CssBaseline enableColorScheme />
-    <App />
-  </ThemeProvider>
+  <StyledEngineProvider injectFirst>
+    <ThemeProvider theme={theme}>
+      <CssBaseline enableColorScheme />
+      <RouterProvider router={router} />
+    </ThemeProvider>
+  </StyledEngineProvider>
 )
